style: move ArgumentList to the libdebug.data package
